requirements.txt
beautifulsoup4
googlesearch-python
googletrans==4.0.0-rc1
python-docx
requests
sumy